Jit64: fselx - Prefer non-AVX sometimes (AVX)
We previously started using VBLENDVPD when AVX is available. This
instruction allows for a separate destination register, which can
sometimes be used to eliminate a move (MOVAPS).

However, VBLENDVPD takes one additional μop compared to BLENDVPD on
Skylake and beyond (according to https://uops.info/table.html).

Rather than unconditionally using VBLENDVPD, we now try to use it only
in the cases where it actually saves a move, and fall back to BLENDVPD
otherwise.