Jit64: subfcx - Optimize constant a
Consider the case where a holds a constant and d == a. Because of the
ReadWrite constraint on this register, the RegCache will allocate a host
register and overwrite it with the constant.

This approach works against us in this case, because we're really trying
to calculate b - a. So we just end up moving the constant that was just
written to the register out of the way again.

By explictly handling this case, we can avoid this.

Before:
41 BE 08 00 00 00    mov         r14d,8
41 8B C6             mov         eax,r14d
44 8B F6             mov         r14d,esi
44 2B F0             sub         r14d,eax

After:
44 8B F6             mov         r14d,esi
41 83 EE 08          sub         r14d,8