Jit_Integer: mulhwx - Improved constant b handling
One-operand IMUL does not accept immediates. We previously used
gpr.UseNoImm(b) to communicate this constraint to the RegCache. The
RegCache then satisfies this constraint by allocating a new register
when dealing with an immediate.

But IMUL expects one of its operands in EAX anyway, so we might as well
move the immediate to EAX instead of allocating a new register to hold
it. Multiplication is commutative, so swapping the operands around gives
us the desired behavior without changing the logic too much.

This particular scenario happens in SoulCalibur II.

Before:
41 BE E8 03 00 00    mov         r14d,3E8h
41 8B C5             mov         eax,r13d
41 F7 EE             imul        r14d
44 8B FA             mov         r15d,edx

After:
B8 E8 03 00 00       mov         eax,3E8h
41 F7 ED             imul        r13d
44 8B F2             mov         r14d,edx