Jit_integer: mulhw - Improve d == a or d == b
When an input register is not in a host register already and overlaps
with the destination register, the RegCache will eagerly generate code
to move the input register's value into a newly allocated host register
for us.

This is a good idea for most use-cases, but not here because one-operand
IMUL is very rigid in where it reads and writes. As a result, we just
end up moving the value around a second time.

To resolve this, we separate our interactions with the RegCache into two
parts; one handling the reads and the multiplication, one to write the
result in the proper location.

Before:
41 BD 67 66 66 66    mov         r13d,66666667h
41 8B C5             mov         eax,r13d
F7 EF                imul        edi
44 8B EA             mov         r13d,edx

After:
B8 67 66 66 66       mov         eax,66666667h
F7 EF                imul        edi
44 8B EA             mov         r13d,edx